# loader: use atomic.Value to store snapshot