--- a/loader/loader_test.go
+++ b/loader/loader_test.go
@@ -190,3 +190,19 @@ func TestDirectoryRefresher(t *testing.T) {
 	snapshot = loader.Snapshot()
 	assert.Equal("hello3", snapshot.Get("file2"))
 }
+
+func BenchmarkSnapsot(b *testing.B) {
+	var ll Loader
+	for i := 0; i < b.N; i++ {
+		ll.Snapshot()
+	}
+}
+
+func BenchmarkSnapsot_Parallel(b *testing.B) {
+	ll := new(Loader)
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			ll.Snapshot()
+		}
+	})
+}