This eliminates the need to use a mutex.

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkSnapsot-12              15.1          2.46          -83.71%
BenchmarkSnapsot_Parallel-12     36.6          0.51          -98.60%
```